Impl `update_config` to support remote update
- Removed `reload_config` b/c it's been deprecated.

There are code patterns like `writer.lock().unwrap_or_else(...)` which
ignore `PoisonError`, assuming that the log writer should not cause a
panic even if another thread has poisoned the lock.

If `log_dir` changes by updating config, logs should be saved in that
directory. This commit implements it by wrapping the file writer.
Alternatively, `tracing_subscriber::reload` could be used, but it cannot
handle cases where the directory is added or removed (as it requires
reassigning the writer itself, which is not possible due to internal
limitations).

Close: #112